added name row
