+++ client/src/components/Pages/OrgDashboard/OrgDashboard.style.js
@@ -3,6 +3,13 @@ import { Link } from 'react-router-dom';
 
 import { size } from '../../../theme';
 
+export const Wrapper = styled.div`
+  width: 100%;
+  margin-top: ${({ mobile }) => mobile && '-2rem'};
+`;
+
+// OLD STYLES
+
 export const PageWrapper = styled.div`
   padding-top: 4rem;
   padding-bottom: 8rem;
