+++ client/src/components/Pages/OrgDashboard/styleProperties.js
@@ -1,0 +1,27 @@
+import * as T from '../../Common/Typography';
+
+export const bottomMargins = {
+  row: {
+    desktop: 5,
+    mobile: 0,
+  },
+  col: {
+    desktop: 0,
+    mobile: 2,
+  },
+  sectionTitle: {
+    desktop: '1rem',
+    mobile: '1rem',
+  },
+};
+
+export const typographies = {
+  headerTitle: {
+    desktop: T.H2,
+    mobile: T.H3,
+  },
+  sectionTitle: {
+    desktop: T.H4C,
+    mobile: T.H6C,
+  },
+};
